Read configuration

+++ software/arduino/BaudrateConverter/BaudrateConverter/functions.h
@@ -1,7 +1,7 @@
 #ifndef _FUNCTIONS_H
 #define _FUNCTIONS_H
 
-const unsigned int get_baud_rate();
+const unsigned long get_baud_rate();
 const bool get_ascii_enabled();
 const int get_mode();
 #endif
